Splitting the common module into 3 separate modules to properly encapsulate service and dao behavior
