--- fwc-common/src/main/java/org/gitter/fwc/services/InstitutionRestController.java
@@ -1,9 +1,0 @@
-package org.gitter.fwc.services;
-
-import java.util.List;
-import java.util.Map;
-
-public interface InstitutionRestController {
-
-    public List<Map<String, String>> findInstitutions(String keyword);
-}
