+++ fwc-dao-spec/.gitignore
@@ -1,0 +1,3 @@
+/target/
+/.classpath
+/.project
