+++ fwc-dao-spec/.settings/org.eclipse.wst.validation.prefs
@@ -1,0 +1,2 @@
+disabled=06target
+eclipse.preferences.version=1
