+++ fwc-dao-spec/pom.xml
@@ -1,0 +1,40 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>org.gitter</groupId>
+		<artifactId>fwc</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>fwc-dao-spec</artifactId>
+	<packaging>jar</packaging>
+
+	<name>Framework Comparison DAO layer specification</name>
+	<description>
+        This project holds dao layer interfaces and exceptions.
+    </description>
+
+	<contributors>
+		<contributor>
+			<name>Jeff Gitter</name>
+			<email>jgitter@nvisia.com</email>
+			<organization>NVisia, LLC</organization>
+			<organizationUrl>www.nvisia.com</organizationUrl>
+		</contributor>
+	</contributors>
+
+	<dependencies>
+		<dependency>
+			<groupId>org.gitter</groupId>
+			<artifactId>fwc-common</artifactId>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.jboss.spec.javax.ejb</groupId>
+			<artifactId>jboss-ejb-api_3.2_spec</artifactId>
+			<scope>provided</scope>
+		</dependency>
+	</dependencies>
+</project>
