+++ fwc-dao-spec/src/main/java/org/gitter/fwc/exception/DataAccessException.java
