+++ fwc-ejb-web/.gitignore
@@ -1,2 +1,3 @@
 /target/
 /.classpath
+/example-generated-service/
