+++ fwc-ejb-web/.settings/org.eclipse.wst.common.component
@@ -4,7 +4,6 @@
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
-        <wb-resource deploy-path="/WEB-INF/classes" source-path="/target/generated-sources"/>
         <dependent-module archiveName="fwc-ejb-services-1.0.0-SNAPSHOT.jar" deploy-path="/WEB-INF/lib" handle="module:/resource/fwc-ejb-services/fwc-ejb-services">
             <dependency-type>uses</dependency-type>
         </dependent-module>
