+++ fwc-service-spec/pom.xml
@@ -8,7 +8,7 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>fwc-common</artifactId>
+    <artifactId>fwc-service-spec</artifactId>
     <packaging>jar</packaging>
 
     <name>Framework Comparison service specification</name>
@@ -26,10 +26,10 @@
     </contributors>
 
     <dependencies>
-        <dependency>
-            <groupId>org.jboss.spec.javax.ejb</groupId>
-            <artifactId>jboss-ejb-api_3.2_spec</artifactId>
-            <scope>provided</scope>
-        </dependency>
+    	<dependency>
+    		<groupId>org.gitter</groupId>
+    		<artifactId>fwc-common</artifactId>
+    		<scope>provided</scope>
+    	</dependency>
     </dependencies>
 </project>
