+++ fwc-spring-services/pom.xml
@@ -31,10 +31,20 @@
         </dependency>
         <dependency>
             <groupId>org.gitter</groupId>
+            <artifactId>fwc-dao-spec</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.gitter</groupId>
+            <artifactId>fwc-service-spec</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <!-- <dependency>
+            <groupId>org.gitter</groupId>
             <artifactId>fwc-dao</artifactId>
             <type>ejb</type>
             <scope>provided</scope>
-        </dependency>
+        </dependency> -->
       <dependency>
         <groupId>org.jboss.spec.javax.servlet</groupId>
         <artifactId>jboss-servlet-api_3.1_spec</artifactId>
