+++ fwc-spring-services/src/main/java/org/gitter/fwc/endpoints/InstitutionSoapServiceImpl.java
@@ -1,0 +1,46 @@
+package org.gitter.fwc.endpoints;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import org.gitter.fwc.exception.FwcServiceException;
+import org.gitter.fwc.services.InstitutionService;
+import org.gitter.fwc.services.InstitutionSoapService;
+import org.gitter.fwc.transfer.FindInstitutionsRequest;
+import org.gitter.fwc.transfer.FindInstitutionsResponse;
+import org.gitter.fwc.transfer.Institution;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.ws.server.endpoint.annotation.Endpoint;
+import org.springframework.ws.server.endpoint.annotation.PayloadRoot;
+import org.springframework.ws.server.endpoint.annotation.RequestPayload;
+import org.springframework.ws.server.endpoint.annotation.ResponsePayload;
+
+@Endpoint
+public class InstitutionSoapServiceImpl implements InstitutionSoapService {
+	private InstitutionService service;
+
+	@Autowired
+	public InstitutionSoapServiceImpl(InstitutionService service) {
+		this.service = service;
+	}
+
+	@Override
+	@PayloadRoot(localPart = FindInstitutionsRequest.NAME, namespace = InstitutionSoapService.TARGET_NAMESPACE)
+	public @ResponsePayload FindInstitutionsResponse findInstitutions(
+			@RequestPayload FindInstitutionsRequest request) throws FwcServiceException {
+		FindInstitutionsResponse response = new FindInstitutionsResponse();
+
+		List<Institution> list = new ArrayList<Institution>();
+		List<Map<String, String>> results = service.findByKeyword(request.getKeyword());
+		for (Map<String, String> result : results) {
+			Institution institution = new Institution();
+			institution.setInstitutionData(result);
+			list.add(institution);
+		}
+
+		response.setInstitutions(list);
+
+		return response;
+	}
+}
